Separating each components to files.
