+++ src/client/components/applayout/header/HeaderRight.js
@@ -1,0 +1,14 @@
+import React from 'react';
+
+class HeaderRight extends React.Component {
+  render() {
+    const customStyle = {
+      maxHeight: '60px'
+    };
+    return (
+      <a className={'pull-left'}><img src={'/static/images/logo.png'} style={customStyle}/></a>
+    );
+  }
+}
+
+export default HeaderRight;
